API: universal support for array types
This change introduces a `glass.array` module which can be imported as
the array namespace:

    import glass.array as xp

By default, that loads the numpy array namespace, so that `xp` is mostly
equivalent to the usual `np`.

However, the `glass.array` module has a function `use` for switching to
a different array namespace internally:

    xp.use('numpy')  # default
    xp.use('cupy')
    xp.use('torch')

This can also be done locally using the context manager-form of the same
command:

    with xp.use('torch'):
        ...

Internally, the `array-api-compat` package is used to load the various
array namespaces into modules `glass._array_api.<name>`.  These modules
can be extended with functionality for GLASS, such as e.g. a spherical
harmonic transform that supports the array type.

The `glass._array_api` module should also be a namespace package, so
that we can have extensions which add support for other array types,
e.g. `glass-jax`.