API(points): support for array inputs
Allow dimensional input to the sampling functions in `glass.points`.
In the language of the `glass.galaxies` module, leading dimensions would
correspond to separate galaxy populations.

Numpy broadcasting rules are applied to the leading dimensions.

Since the sampling returns a potentially different number of points for
each extra dimension, the return value for dimensional input is a 1-D
list of points, corresponding to the flattened leading dimensions.

To simplify working with dimensional output, the sampling functions now
always return a third value `counts` along with longitudes and
latitudes, which is either an int or a list of ints.  Because the return
values are different, this is a breaking change.

BREAKING CHANGE: Position sampling returns counts alongside points.