gh-231: add `ruff` linting to `pre-commit` (#232)
Add `ruff` linting and turn on as many rules as possible. However, I
will turn off many and not spend too much time fixing them. In a perfect
world, we will try and fix as many of these as possible - as they will
generally help with maintenance in the long term.

Closes: #231
Refs: #170
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nicolas Tessore <n.tessore@ucl.ac.uk>
Co-authored-by: Saransh Chopra <saransh0701@gmail.com>